django-archive/requirements.txt
django-2.2
